# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CMS CRUD Restful API using Sequelize and Swagger",
    "description": "CMS CRUD Restful API with authentication and authorization. Using PostgreSQL as DBMS and Sequelize as its ORM. Documentation is served using Swagger.",
    "contact": {
      "name": "Doni Rahma Tiana",
      "email": "donirahmatiana@gmail.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://mighty-eyrie-29994.herokuapp.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/articles": {
      "get": {
        "tags": ["ARTICLES"],
        "summary": "Read Articles",
        "description": "Get all articles",
        "operationId": "ReadArticles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get articles success"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": ["ARTICLES"],
        "summary": "Create Article",
        "description": "Create article",
        "operationId": "CreateArticle",
        "parameters": [],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleRequest"
              },
              "example": {
                "title": "Article Title",
                "content": "Article content"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Article created"
          },
          "400": {
            "description": "Bad input"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an user"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      }
    },
    "/api/articles/{article_id}": {
      "get": {
        "tags": ["ARTICLES"],
        "summary": "Read Article",
        "description": "Get article by id",
        "operationId": "ReadArticle",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get article success"
          },
          "400": {
            "description": "Param article_id is not exists"
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": ["ARTICLES"],
        "summary": "Update Article",
        "description": "Update article by article id",
        "operationId": "UpdateArticle",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleRequest"
              },
              "example": {
                "title": "Updated Article Title",
                "content": "Updated article content"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Article updated"
          },
          "400": {
            "description": "Bad input or param article_id not exists"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not article owner"
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": ["ARTICLES"],
        "summary": "Delete Article",
        "description": "Delete article by id",
        "operationId": "DeleteArticle",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Article deleted"
          },
          "400": {
            "description": "Param article_id is not exists"
          },
          "403": {
            "description": "No token provided or not article owner"
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/users/comments/{comment_id}": {
      "delete": {
        "tags": ["USER COMMENT"],
        "summary": "Delete Comment",
        "description": "Delete a comment",
        "operationId": "DeleteComment",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "description": "comment id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted"
          },
          "400": {
            "description": "Param comment_id is not exists"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided, not an user, or not comment owner"
          },
          "404": {
            "description": "Comment not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      }
    },
    "/api/articles/{article_id}/comments": {
      "get": {
        "tags": ["COMMENT"],
        "summary": "Read Comments",
        "description": "Get comments by article id",
        "operationId": "ReadComments",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get comments of an user article success"
          },
          "400": {
            "description": "Param article_id is not exists"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an user"
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": ["ADMIN USER"],
        "summary": "Read Users",
        "description": "Admin get all user",
        "operationId": "ReadUsers",
        "parameters": [],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get users success"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an admin"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": ["ADMIN USER"],
        "summary": "Create User",
        "description": "Admin create new user",
        "operationId": "CreateUser",
        "parameters": [],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              },
              "example": {
                "name": "Doni created",
                "username": "DoniCreated",
                "email": "doniCreated@gmail.com",
                "password": "12345678"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Bad input or duplicate account found"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an admin"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      }
    },
    "/api/admin/users/{user_id}": {
      "get": {
        "tags": ["ADMIN USER"],
        "summary": "Read User",
        "description": "Admin get user by Id",
        "operationId": "ReadUser",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "description": "user id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get user success"
          },
          "400": {
            "description": "Param user_id is not exists"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an admin"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": ["ADMIN USER"],
        "summary": "Update User",
        "description": "Admin update user",
        "operationId": "UpdateUser",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "description": "user id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              },
              "example": {
                "name": "User updated",
                "username": "UserUpdated",
                "email": "userUpdated@gmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated"
          },
          "400": {
            "description": "Bad input, ing to modify superadmining to modify superadmin account, or duplicate account found"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an admin"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": ["ADMIN USER"],
        "summary": "Delete User",
        "description": "Admin delete user",
        "operationId": "DeleteUser",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "description": "user id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "400": {
            "description": "Bad input, param user_id is not exists, or intending to modify superadmin"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an admin"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/admin/articles/{article_id}": {
      "get": {
        "tags": ["ADMIN ARTICLES"],
        "summary": "Read Article",
        "description": "Admin get article by id",
        "operationId": "GetReadArticle",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get article success"
          },
          "400": {
            "description": "Param article_id is not exists"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an admin"
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": ["ADMIN ARTICLES"],
        "summary": "Delete article",
        "description": "Admin delete article by id",
        "operationId": "Deletearticle",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "httpBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Article deleted"
          },
          "400": {
            "description": "Param article_id is not exists"
          },
          "401": {
            "description": "Authentication failed"
          },
          "403": {
            "description": "No token provided or not an admin"
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateArticleRequest": {
        "title": "CreateArticleRequest",
        "required": ["title", "content"],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "example": {
          "title": "Article Title",
          "content": "Article content"
        }
      },
      "UpdateArticleRequest": {
        "title": "UpdateArticleRequest",
        "required": [],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "example": {
          "title": "Updated Article Title",
          "content": "Updated article content"
        }
      },
...
      "CreateUserRequest": {
        "title": "CreateUserRequest",
        "required": ["name", "username", "email", "password"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "name": "Doni created",
          "username": "DoniCreated",
          "email": "doniCreated@gmail.com",
          "password": "12345678"
        }
      },
      "UpdateUserRequest": {
        "title": "UpdateUserRequest",
        "required": [],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "example": {
          "name": "User updated",
          "username": "UserUpdated",
          "email": "userUpdated@gmail.com"
        }
      }
    },
    "securitySchemes": {
      "httpBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}